Fix the Usernames with NaN elo ladder bug
Just a quick fix to the ladder bug where Usernames with NaN elo start popping up. Got suggested this by a showdown staff member for a server I work on so thought you would appreciate it since you have 13 of the top spots filled with them on the ou ladder.